refactor buffers
This splits buffers.py, which contained all buffer classes, into several smaller files.
issue #1226
